[feature] set scan Filter sugar generation related by mode type
